Split the 00_dump into base + django + OI base
The dump file was mainly aiming at converting existing database and was creating the DB as it was deployed, with some migrations.
I split this into a main dump in the old format and the remaining tables in a separate file.
More fixes will now be needed for a conversion of existing DB. Coming soon